requirements.txt
astroid==1.6.5
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-ckeditor==5.5.0
django-colorfield==0.1.14
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-jsoneditor==0.0.12
django-modeltranslation==0.13b1
django-mptt==0.9.0
django-qr-code==1.0.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
MonthDelta==0.9.1
packaging==17.1
Pillow==5.1.0
psycopg2==2.7.5
pycodestyle==2.4.0
pylint==1.9.2
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
qrcode==6.0
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
wrapt==1.10.11
